LayoutTests/imported/w3c/web-platform-tests/custom-elements/cross-realm-callback-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS
CONSOLE MESSAGE: Error: PASS
CONSOLE MESSAGE: Error: PASS
CONSOLE MESSAGE: Error: PASS
CONSOLE MESSAGE: Error: PASS


PASS constructor
PASS connectedCallback
PASS disconnectedCallback
PASS attributeChangedCallback
PASS adoptedCallback
